(PC-23615) refactor(gtl): apply feedbacks
